pyproject.toml
[project]
name = "particulars-analysis"
version = "1.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.*"
dependencies = [
    "astral==3.2",
    "contourpy==1.3.2",
    "cycler==0.12.1",
    "fonttools==4.58.4",
    "gitdb==4.0.12",
    "gitpython==3.1.44",
    "kiwisolver==1.4.8",
    "matplotlib==3.10.3",
    "numpy==1.26.4",
    "packaging==25.0",
    "pandas==2.3.0",
    "pillow==11.2.1",
    "pyparsing==3.2.3",
    "pyside6==6.9.1",
    "pyside6-addons==6.9.1",
    "pyside6-essentials==6.9.1",
    "python-dateutil==2.9.0.post0",
    "pytz==2024.2",
    "scipy==1.16.0",
    "shiboken6==6.9.1",
    "six==1.17.0",
    "smmap==5.0.2",
    "sun-position-calculator==1.0.0",
    "tzdata==2024.2",
]
